Unify FP types in scaling code and avoid type-casting back and forth

+++ core_lib/src/managers/viewmanager.h
@@ -62,8 +62,8 @@ public:
     float rotation();
     void rotate(float degree);
 
-    float scaling();
-    void scale(float scaleValue, QPointF offset = {0, 0});
+    qreal scaling();
+    void scale(qreal scaleValue, QPointF offset = {0, 0});
     void scaleUp();
     void scaleDown();
     void scale400();
